LPS-161984 Extract AttachmentProperties from EditObjectField to its own component
