+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/AttachmentProperties.tsx
@@ -1,0 +1,92 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import ClayForm from '@clayui/form';
+import {Input} from '@liferay/object-js-components-web';
+import {sub} from 'frontend-js-web';
+import React from 'react';
+
+import {normalizeFieldSettings} from '../../utils/fieldSettings';
+import {ObjectFieldErrors} from '../ObjectFieldFormBase';
+
+export function AttachmentProperties({
+	errors,
+	objectFieldSettings,
+	onSettingsChange,
+}: IAttachmentPropertiesProps) {
+	const settings = normalizeFieldSettings(objectFieldSettings);
+
+	return (
+		<>
+			<ClayForm.Group>
+				{settings.showFilesInDocumentsAndMedia && (
+					<Input
+						error={errors.storageDLFolderPath}
+						feedbackMessage={sub(
+							Liferay.Language.get(
+								'input-the-path-of-the-chosen-folder-in-documents-and-media-an-example-of-a-valid-path-is-x'
+							),
+							'/myDocumentsAndMediaFolder'
+						)}
+						label={Liferay.Language.get('storage-folder')}
+						maxLength={255}
+						onChange={({target: {value}}) =>
+							onSettingsChange({
+								name: 'storageDLFolderPath',
+								value,
+							})
+						}
+						required
+						value={settings.storageDLFolderPath as string}
+					/>
+				)}
+			</ClayForm.Group>
+			<Input
+				component="textarea"
+				error={errors.acceptedFileExtensions}
+				feedbackMessage={Liferay.Language.get(
+					'enter-the-list-of-file-extensions-users-can-upload-use-commas-to-separate-extensions'
+				)}
+				label={Liferay.Language.get('accepted-file-extensions')}
+				onChange={({target: {value}}) =>
+					onSettingsChange({name: 'acceptedFileExtensions', value})
+				}
+				required
+				value={settings.acceptedFileExtensions as string}
+			/>
+
+			<Input
+				error={errors.maximumFileSize}
+				feedbackMessage={Liferay.Language.get('maximum-file-size-help')}
+				label={Liferay.Language.get('maximum-file-size')}
+				min={0}
+				onChange={({target: {value}}) =>
+					onSettingsChange({
+						name: 'maximumFileSize',
+						value: value && Number(value),
+					})
+				}
+				required
+				type="number"
+				value={settings.maximumFileSize as number}
+			/>
+		</>
+	);
+}
+
+interface IAttachmentPropertiesProps {
+	errors: ObjectFieldErrors;
+	objectFieldSettings: ObjectFieldSetting[];
+	onSettingsChange: (setting: ObjectFieldSetting) => void;
+}
