+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/EditObjectField.tsx
@@ -12,7 +12,6 @@
  * details.
  */
 
-import ClayForm from '@clayui/form';
 import {useModal} from '@clayui/modal';
 import {
 	API,
@@ -25,22 +24,16 @@ import {
 	openToast,
 	saveAndReload,
 } from '@liferay/object-js-components-web';
-import {sub} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
-import {
-	normalizeFieldSettings,
-	updateFieldSettings,
-} from '../../utils/fieldSettings';
+import {updateFieldSettings} from '../../utils/fieldSettings';
 import {
 	FilterErrors,
 	FilterValidation,
 	ModalAddFilter,
 } from '../ModalAddFilter';
-import ObjectFieldFormBase, {
-	ObjectFieldErrors,
-	useObjectFieldForm,
-} from '../ObjectFieldFormBase';
+import ObjectFieldFormBase, {useObjectFieldForm} from '../ObjectFieldFormBase';
+import {AttachmentProperties} from './AttachmentProperties';
 import {MaxLengthProperties} from './MaxLengthProperties';
 import {SearchableContainer} from './SearchableContainer';
 
@@ -646,71 +639,6 @@ export default function EditObjectField({
 	);
 }
 
-function AttachmentProperties({
-	errors,
-	objectFieldSettings,
-	onSettingsChange,
-}: IAttachmentPropertiesProps) {
-	const settings = normalizeFieldSettings(objectFieldSettings);
-
-	return (
-		<>
-			<ClayForm.Group>
-				{settings.showFilesInDocumentsAndMedia && (
-					<Input
-						error={errors.storageDLFolderPath}
-						feedbackMessage={sub(
-							Liferay.Language.get(
-								'input-the-path-of-the-chosen-folder-in-documents-and-media-an-example-of-a-valid-path-is-x'
-							),
-							'/myDocumentsAndMediaFolder'
-						)}
-						label={Liferay.Language.get('storage-folder')}
-						maxLength={255}
-						onChange={({target: {value}}) =>
-							onSettingsChange({
-								name: 'storageDLFolderPath',
-								value,
-							})
-						}
-						required
-						value={settings.storageDLFolderPath as string}
-					/>
-				)}
-			</ClayForm.Group>
-			<Input
-				component="textarea"
-				error={errors.acceptedFileExtensions}
-				feedbackMessage={Liferay.Language.get(
-					'enter-the-list-of-file-extensions-users-can-upload-use-commas-to-separate-extensions'
-				)}
-				label={Liferay.Language.get('accepted-file-extensions')}
-				onChange={({target: {value}}) =>
-					onSettingsChange({name: 'acceptedFileExtensions', value})
-				}
-				required
-				value={settings.acceptedFileExtensions as string}
-			/>
-
-			<Input
-				error={errors.maximumFileSize}
-				feedbackMessage={Liferay.Language.get('maximum-file-size-help')}
-				label={Liferay.Language.get('maximum-file-size')}
-				min={0}
-				onChange={({target: {value}}) =>
-					onSettingsChange({
-						name: 'maximumFileSize',
-						value: value && Number(value),
-					})
-				}
-				required
-				type="number"
-				value={settings.maximumFileSize as number}
-			/>
-		</>
-	);
-}
-
 interface AggregationFilters {
 	defaultSort?: boolean;
 	fieldLabel?: string;
@@ -729,12 +657,6 @@ interface AggregationFilters {
 interface IItem extends LabelValueObject {
 	checked?: boolean;
 }
-interface IAttachmentPropertiesProps {
-	errors: ObjectFieldErrors;
-	objectFieldSettings: ObjectFieldSetting[];
-	onSettingsChange: (setting: ObjectFieldSetting) => void;
-}
-
 interface IProps {
 	filterOperators: TFilterOperators;
 	forbiddenChars: string[];
